requirements.txt
absl-py==1.4.0            # via tensorboard, tensorflow-cpu
astunparse==1.6.3         # via tensorflow-cpu
cachetools==5.2.1         # via google-auth
certifi==2022.12.7        # via requests
charset-normalizer==3.0.1  # via requests
flatbuffers==23.1.4       # via tensorflow-cpu
gast==0.4.0               # via tensorflow-cpu
google-auth==2.16.0       # via google-auth-oauthlib, tensorboard
google-auth-oauthlib==0.4.6  # via tensorboard
google-pasta==0.2.0       # via tensorflow-cpu
grpcio==1.51.1            # via tensorboard, tensorflow-cpu
h5py==3.7.0               # via tensorflow-cpu
idna==3.4                 # via requests
keras==2.11.0             # via tensorflow-cpu
libclang==15.0.6.1        # via tensorflow-cpu
markdown==3.4.1           # via tensorboard
markupsafe==2.1.2         # via werkzeug
numpy==1.24.1             # via h5py, opt-einsum, tensorboard, tensorflow-cpu
oauthlib==3.2.2           # via requests-oauthlib
opt-einsum==3.3.0         # via tensorflow-cpu
packaging==23.0           # via tensorflow-cpu
protobuf==3.19.6          # via tensorboard, tensorflow-cpu
pyasn1==0.4.8             # via pyasn1-modules, rsa
pyasn1-modules==0.2.8     # via google-auth
pygame==2.1.2             # via -r requirements.in
requests==2.28.2          # via requests-oauthlib, tensorboard
requests-oauthlib==1.3.1  # via google-auth-oauthlib
rsa==4.9                  # via google-auth
six==1.16.0               # via astunparse, google-auth, google-pasta, tensorflow-cpu
tensorboard==2.11.2       # via tensorflow-cpu
tensorboard-data-server==0.6.1  # via tensorboard
tensorboard-plugin-wit==1.8.1  # via tensorboard
tensorflow-cpu==2.11.0    # via -r requirements.in
tensorflow-estimator==2.11.0  # via tensorflow-cpu
tensorflow-io-gcs-filesystem==0.29.0  # via tensorflow-cpu
termcolor==2.2.0          # via tensorflow-cpu
typing-extensions==4.4.0  # via tensorflow-cpu
urllib3==1.26.14          # via requests
werkzeug==2.2.2           # via tensorboard
wheel==0.38.4             # via astunparse, tensorboard
wrapt==1.14.1             # via tensorflow-cpu

# The following packages are considered to be unsafe in a requirements file:
# setuptools
